Add ability to specify user and group for files installed by package.
The use case is I often want the files on disk to be owned by 'root'
so that even if my service (running as a non-root user named
after the virtualenv) is compromised, the attacker can't modify
the on-disk representation of the program.

New tests for this feature are included. I've confirmed that the tests
fail when the code change is not in place.